[Mod/#25] user api 분리
