+++ app/src/main/java/org/sopt/carrot/data/service/ProductService.kt
@@ -3,17 +3,8 @@ package org.sopt.carrot.data.service
 import org.sopt.carrot.data.model.response.BaseResponse
 import org.sopt.carrot.data.model.response.ResponseProductInfoDto
 import org.sopt.carrot.data.model.response.ResponseRelatedProductsDto
-import org.sopt.carrot.data.model.response.ResponseSearchDto
 import retrofit2.http.GET
 import retrofit2.http.Path
-import retrofit2.http.Query
-
-interface ProductApiService {
-interface ProductService {
-    @GET("/api/search")
-    suspend fun searchProducts(
-        @Query("keyword") keyword: String
-    ): BaseResponse<ResponseSearchDto>
 
 interface ProductApiService {
     @GET("/api/products/{productId}")
