+++ app/src/main/java/org/sopt/carrot/domain/model/ProductDetail.kt
@@ -1,12 +1,5 @@
 package org.sopt.carrot.domain.model
 
-data class UserDetailModel(
-    val userId: Long,
-    val nickname: String,
-    val profileImage: String? = null,
-    val address: String
-)
-
 data class ProductDetailModel(
     val productId: Long,
     val productImage: String,
