+++ app/src/main/java/org/sopt/carrot/domain/model/User.kt
@@ -1,0 +1,8 @@
+package org.sopt.carrot.domain.model
+
+data class UserDetailModel(
+    val userId: Long,
+    val nickname: String,
+    val profileImage: String? = null,
+    val address: String
+)
